Change transactions.schema.amount to float

+++ db/migrations/20181020100815_transactions.js
@@ -2,7 +2,7 @@ exports.up = knex => knex.schema.createTable('transactions', (table) => {
   table.increments();
   table.integer('account_id').references('accounts.id').onDelete('CASCADE');
   table.string('title').notNullable();
-  table.decimal('amount').notNullable();
+  table.float('amount').notNullable();
   table.boolean('pending').defaultTo(true);
   table.timestamps(true, true);
 });
